requirements.txt
boto==2.48.0
boto3==1.6.15
botocore==1.9.15
bz2file==0.98
certifi==2018.1.18
chardet==3.0.4
docutils==0.14
gensim==3.4.0
idna==2.6
Janome==0.3.6
jmespath==0.9.3
mecab-python3==0.7
numpy==1.14.2
python-dateutil==2.6.1
requests==2.18.4
s3transfer==0.1.13
scipy==1.0.0
six==1.11.0
smart-open==1.5.7
urllib3==1.22
